% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mercury.R
\name{remove_html}
\alias{remove_html}
\title{Removes html}
\usage{
remove_html(strings, trim = TRUE)
}
\arguments{
\item{strings}{the string(s) you want to clean}

\item{trim}{should the string be trimmed or not}
}
\value{
a string
}
\description{
The function uses tools from the rvest and xml2 packages to clean up the HTML
and turning it into proper text.
}
\examples{
\dontrun{
# First get api key here: https://mercury.postlight.com/web-parser/

# Then run the code below replacing the X's wih your api key.
url <- "https://trackchanges.postlight.com/building-awesome-cms-f034344d8ed"
my_data <- web_parser(page_urls = url,
                      api_key = XXXXXXXXXXXXXXXXXXXXXXX)

# With html formatting:
my_data$content

# Now remove it:
my_data$content <- remove_html(my_data$content)

# Without html formatting:
my_data$content
}
}
