% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mercury.R
\name{web_parser}
\alias{web_parser}
\title{web_parser}
\source{
\url{https://mercury.postlight.com/web-parser/}
}
\usage{
web_parser(page_urls, api_key)
}
\arguments{
\item{page_urls}{One or more urls to be parsed}

\item{api_key}{Key for the API}
}
\value{
a tibble
}
\description{
With just one API request, Mercury takes any web article and returns only the
relevant content — headline, author, body text, relevant images and more —
free from any clutter. It’s reliable, easy-to-use and free.
}
\examples{
\dontrun{
# First get api key here: https://mercury.postlight.com/web-parser/

# Then run the code below replacing the X's wih your api key.

web_parser(page_urls = "https://trackchanges.postlight.com/building-awesome-cms-f034344d8ed",
           api_key = XXXXXXXXXXXXXXXXXXXXXXX)
}
}
